Remove extra arg from `create_achievement`
Following up on some thoughts that I had in the message for commit `b4d4e976`,
I am removing any positional arguments to the `create_achievement`
function besides the `attrs`.

it is now a unary function, and it makes no expectations about what sort
of map is passed in (i.e. no pattern matching is done), pushing all of
the work of validation and error message handling for bad input onto the
changeset, which is where I think that burden should be placed most of
the time.