Add "streaks" for achievements
This relies on a new DB column that allows us to treat achievement
streaks like a flat tree; each "streak" is comprised of achievements
whose `:streak_start` is all the same date and whose `:activity_id` is
the same.

When a new `%Achievement{}` is created there is an extra DB call now to
gather any achievements for the same activity from yesterday (time zones
will quickly cause problems here, but hey, it's a hobby project) and
find the one with the oldest `:streak_start`, and use that as the
`:streak_start` for the new achievement.

This means that looking up the streak (which will be done much more than
creating new achievements) will be a very cheap operation: if you have
the achievement data on hand, it's just some date-math, and if you don't
have it then it's a single DB call |> date-math

**disclaimer**

one thing that I am not happy with is how I had to "weaken" my
`%Achievement{}` changeset so that I could manipulate the
`:streak_start` and `:inserted_at` timestamps to test the timing logic.
I am not sure how to test this in a more hygienic way besides breaking
the logic for "do these achievements constitute a streak" out into its
own function. At present, since that logic is not needed anywhere else
in the app, I don't want it to be exposed, but I don't know if that
should outweigh the new weakness I have introduced to my DB change
policy