Use "card" for `Activity` on "show" page, broadcast achievement creation
There are a large number of changes in this commit, strung together over
about a week of picking up my laptop every now and again to hack at this
in between all sorts of other things. Admittedly, it's not very cohesive
and there are some clear rough edges in the implementation now.

I am starting to feel a good deal of pain around the "phoenix context"
abstraction for non-trivial things like the data setup required to
render all of the info that I want for an `Activity`'s "show" page. I
think that my next major action is to clean up (and possibly remove most
of) my context functions, and allow what is currently the client code of
those functions  much more control over their use of the database, since
_really_ only the client code knows what kind of data it needs, and is
thus in the best position to form the appropriate queries (within
reason).