Add new function for calculating streaks
This new function does not rely on any special database columns like the
previous implementation did, making it more flexible.

I elected to make this a function in the `Activity` module and not as
some reduction over `Achievement`s in that module because it feels like
a question to me that is _only relevant_ when talking about a given
activity, and abstracting the details of exactly how the calculation is
done for a given activity feels to me like it "deepens" the `Activity`
module.