chore - add `credo` and `typed_ecto_schema`
Adding `credo` pointed out that using the `%Struct{}` syntax
inside of a `@spec` was creating compile-time dependencies between my
modules.

I alleviated that by using `typed_ecto_schema` to generate a
`Struct.t()` type for each of my schema modules so that they can easily
be used in `@spec`s. Using the library lets me describe the schema once,
in the schema definition (i.e. the thing that really counts) and then
get a `@type` spec out of that "for free"