Add `AchievementRelay` GenServer && add `Mox`
The primary goal of this commit was to put something in place to resolve
issue #1 by creating a mechanism to disseminate events for certain
topics.

Along the way I realized that this could be very difficult to unti test in a
consistent way, and decided to use `Mox` based on my recollection of its
use in the Jeff Matthias and Andrea Leopardi "Testing Elixir" book.
Defining a behaviour (required to use `Mox`) felt fairly natural for
this implementation/ plan.

My setup for the mocking, stubbing, and expectation in the test itself
come from reading the official Mox docs on hex.

Resolves #1