Convert all datetimes to use UTC with microsecond precision
`usec` -> microsecond (the u is for the greek "mu" character, which is
the common prefix for "micro"/ 10^(-6))

I generally think it is best for all stored datetimes to use UTC, and I
was adhering to this up until now by always using naive datetimes and
only ever generating them with "utc" functions. however this explicit
conversion removes any uncertainty from the mix, and is actually what
the ecto team _wishes_ they had done as the default.

I used the following resources in researching and implementing these
changes:
  - hex docs, extensively
  - https://elixirforum.com/t/why-use-utc-datetime-over-naive-datetime-for-ecto/32532/5
    - excellent discussion with nuggets of wisdom from Wojtek and
      "LostKobrakai" (Benjamin Milde)
  - http://www.creativedeletion.com/2019/06/17/utc-timestamps-in-ecto.html