Add timezone localization for streaks, points, and displayed dates
I also made some minor changes to the displayed values, particulaly for
recent achievements because I feel that they do not need very much info
since they are recent (not necessary to give the year, for instance)

I moved a number of display concerns into the display code itself. It
feels to me that that is the best place to do formatting of dates, etc.,
instead of handling it in the `handle_params` callback or similar. If
you wanna know how something is being formatted, you go to the same code
that puts it on the screen