Add a credo config file, modify default nesting rule
I received a credo refactor warning about overly nested code. I
personally think that nesting does not indicate poorly factored code
in-and-of-itself, so I at least bumped up the warning threshold to > 3.
In this particular case, the code that I was being warned about is in
the `active_streak_start/1` function, but I believe that all of that
code is very tightly related and is not general purpose enough to
extract and place somewhere else; it would only have the one callsite.

My opinion has recently been changed in this regard thanks to [A
Philosophy of Software Design](https://web.stanford.edu/~ouster/cgi-bin/book.php)