Add range info for streak where applicable
This is a **breaking** change for the `Activities.active_streak_start/1`
function (not only because I changed the name of it!)

The return type has changed to reflect the different cases that can
happen for a streak _and_ return the most information possible to the
caller, instead of requiring more calls/ work on their end in the event
that they need to work out the most recent achievement in the streak

I am struggling with designing these APIs, as I'm not sure how to keep
them "general purpose" but still hide information like the user's
timezone from the callers.

Technically, I have just complicated this one function a bit more for
some functionality that was already available, but would necessitate
another function call (`list_achievements_since/3`)

Going to keep working through these issues as they come up, and spend
some more time thinking about the design of the whole system.